Move duplicate OrientationAnalysis Eigen matrix code to utility file for reusability
